Added $real, $proxy, $wrapper, $context as injectables to the template function. They are also defined directly on the $context object that can be injected into the template function.
Instead of returning an object the template function now should inject
the $context object and edit it.

If you wish you can inject them all the way down to the object to being wrapped to
give it access to it's proxy or wrapper, and the entire context.
This would be useful in the case of an object that has methods that will return a new
 wrapper for itself with a different context.

Updated documentation.